chat_with_postgre/requirements.txt
anthropic==0.42.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
